requirements.txt
Django==3.2.6
celery==5.1.2
drf_yasg==1.20.0
requests==2.26.0
djangorestframework==3.12.4
